+++ BusinessCape/DTOs/SupplyCategory/SupplyCategoryUpdateDto.cs
@@ -2,7 +2,7 @@
 {
     public class SupplyCategoryUpdateDto
     {
-        public long IdSupplyCategory { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; } = null!;
         public string Description { get; set; } = null!;
     }