+++ DataCape/Models/ClientModel.cs
@@ -7,18 +7,18 @@ namespace DataCape.Models
     {
         public ClientModel()
         {
-            QuotationClients = new HashSet<QuotationClientModel>();
+            QuotationClients = new HashSet<QuotationClient>();
         }
 
-        public long IdClient { get; set; }
-        public bool? StatedAt { get; set; }
-        public string? Name { get; set; }
+        public long Id { get; set; }
+        public string Name { get; set; } = null!;
         public string? Phone { get; set; }
-        public string? Email { get; set; }
-        public string? Center { get; set; }
-        public string? Area { get; set; }
-        public string? Regional { get; set; }
+        public string Email { get; set; } = null!;
+        public string Center { get; set; } = null!;
+        public string Area { get; set; } = null!;
+        public string Regional { get; set; } = null!;
+        public bool? StatedAt { get; set; }
 
-        public virtual ICollection<QuotationClientModel> QuotationClients { get; set; }
+        public virtual ICollection<QuotationClient> QuotationClients { get; set; }
     }
 }