+++ DataCape/Models/GrammageCaliberModel.cs
@@ -3,10 +3,11 @@ using System.Collections.Generic;
 
 namespace DataCape.Models
 {
-    public partial class FinishModel
+    public partial class GrammageCaliberModel
     {
         public long Id { get; set; }
-        public string Name { get; set; } = null!;
+        public string Type { get; set; } = null!;
+        public string? Name { get; set; }
         public bool? StatedAt { get; set; }
     }
 }