+++ DataCape/Models/ImpositionPlanchModel.cs
@@ -3,10 +3,11 @@ using System.Collections.Generic;
 
 namespace DataCape.Models
 {
-    public partial class FinishModel
+    public partial class ImpositionPlanchModel
     {
         public long Id { get; set; }
         public string Name { get; set; } = null!;
+        public byte[]? Scheme { get; set; }
         public bool? StatedAt { get; set; }
     }
 }