+++ DataCape/Models/MachineModel.cs
@@ -5,13 +5,7 @@ namespace DataCape.Models
 {
     public partial class MachineModel
     {
-        public MachineModel()
-        {
-            QuotationClients = new HashSet<QuotationClientModel>();
-        }
-
-        public long IdMachine { get; set; }
-        public bool? StatedAt { get; set; }
+        public long Id { get; set; }
         public string? Name { get; set; }
         public decimal? MinimumHeight { get; set; }
         public decimal? MinimumWidth { get; set; }
@@ -19,7 +13,6 @@ namespace DataCape.Models
         public decimal? MaximumWidth { get; set; }
         public decimal? CostByUnit { get; set; }
         public decimal? CostByHour { get; set; }
-
-        public virtual ICollection<QuotationClientModel> QuotationClients { get; set; }
+        public bool? StatedAt { get; set; }
     }
 }