+++ DataCape/Models/QuotationProviderModel.cs
@@ -5,13 +5,13 @@ namespace DataCape.Models
 {
     public partial class QuotationProviderModel
     {
-        public long IdQuotationProvider { get; set; }
-        public DateTime? Date { get; set; }
+        public long Id { get; set; }
+        public DateTime? QuotationDate { get; set; }
         public byte[]? QuotationFile { get; set; }
         public double? FullValue { get; set; }
-        public long? IdProvider { get; set; }
+        public long? ProviderId { get; set; }
         public bool? StatedAt { get; set; }
 
-        public virtual ProviderModel? IdProviderNavigation { get; set; }
+        public virtual ProviderModel? Provider { get; set; }
     }
 }