+++ DataCape/Models/SupplyModel.cs
@@ -10,20 +10,19 @@ namespace DataCape.Models
             SupplyDetails = new HashSet<SupplyDetailModel>();
         }
 
-        public long IdSupply { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; } = null!;
-        public long IdUnitMeasure { get; set; }
         public string? DangerIndicators { get; set; }
         public string? UseInstructions { get; set; }
         public string? Advices { get; set; }
         public int SupplyType { get; set; }
         public int SortingWord { get; set; }
-        public int Quantity { get; set; }
+        public int? Quantity { get; set; }
         public decimal? AverageCost { get; set; }
-        public long IdWarehouse { get; set; }
+        public long? WarehouseId { get; set; }
         public bool? StatedAt { get; set; }
 
-        public virtual WarehouseModel IdWarehouseNavigation { get; set; } = null!;
+        public virtual WarehouseModel? Warehouse { get; set; }
         public virtual ICollection<SupplyDetailModel> SupplyDetails { get; set; }
     }
 }