+++ DataCape/Models/SupplyPictogramModel.cs
@@ -5,9 +5,11 @@ namespace DataCape.Models
 {
     public partial class SupplyPictogramModel
     {
-        public long IdPictogram { get; set; }
+        public long Id { get; set; }
+        public string Code { get; set; } = null!;
         public string Name { get; set; } = null!;
-        public byte[] Pictogram { get; set; } = null!;
+        public string? Description { get; set; }
+        public byte[]? PictogramFile { get; set; }
         public bool? StatedAt { get; set; }
     }
 }