+++ DataCape/Models/SupplyXProductModel.cs
@@ -5,10 +5,10 @@ namespace DataCape.Models
 {
     public partial class SupplyXProductModel
     {
-        public long IdSupply { get; set; }
-        public long IdProduct { get; set; }
+        public long? ProductId { get; set; }
+        public long? SupplyId { get; set; }
 
-        public virtual ProductModel IdProductNavigation { get; set; } = null!;
-        public virtual SupplyModel IdSupplyNavigation { get; set; } = null!;
+        public virtual ProductModel? Product { get; set; }
+        public virtual SupplyModel? Supply { get; set; }
     }
 }