+++ DataCape/Models/WarehouseTypeModel.cs
@@ -5,11 +5,16 @@ namespace DataCape.Models
 {
     public partial class WarehouseTypeModel
     {
-        public long IdTypeWarehouse { get; set; }
-        public bool? StatedAt { get; set; }
-        public string? Nametype { get; set; }
-        public string? Description { get; set; }
+        public WarehouseTypeModel()
+        {
+            Warehouses = new HashSet<WarehouseModel>();
+        }
 
-        public virtual WarehouseModel? Warehouse { get; set; }
+        public long Id { get; set; }
+        public string Name { get; set; } = null!;
+        public string? Description { get; set; }
+        public bool? StatedAt { get; set; }
+
+        public virtual ICollection<WarehouseModel> Warehouses { get; set; }
     }
 }